Separated create/start for all high-level services, and some internal services. There are still internal services that are combined.   These could eventually be turned into simple S4U_Daemon, but they use the setProperty/getProperty that is a Service thing. So eventally, for consistency we could separate their create/start as well.
Modified the run_batsched() method to make it more robust, especially
when batsched is not installed!

+++ src/wrench/simgrid_S4U_util/S4U_Mailbox.cpp
@@ -161,7 +161,7 @@ namespace wrench {
       simgrid::s4u::MailboxPtr mailbox = simgrid::s4u::Mailbox::byName(mailbox_name);
 
       try {
-        mailbox->put_init(msg, msg->payload)->detach();
+        mailbox->put_init(msg, (uint64_t) msg->payload)->detach();
       } catch (xbt_ex &e) {
         if ((e.category == network_error) || (e.category == timeout_error)) {
           WRENCH_INFO("Network error while doing a dputMessage()");
